Split into separate files
OpenAPI has support for multiple files using `$ref`. This uses that to split off
the partner API into separate files. To make it easy to find the files I'm
nameing them exactly like the `path` in the main swagger file.

There's one catch though. The cloudbuild has to be updated to include all the
new files, and to not add to the pile of one-liners I'm moving the bash script
into its own file. This has a caveat in that Cloud Build uses _substitutions_
instead of environment variables. The fix is easy enough, pass BRANCH_NAME as an
explicit argument to the script.